requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
modelscope==1.17.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.3
openai==1.38.0
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0+cu124
tqdm==4.66.5
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2

pymilvus~=2.4.5